Switched from using pynput to using xdotool.
This has a drawback, but also a plus. The drawback is that hatel can only be
run on systems with Xorg running. The plus is that keys are more reliable now.
I could not get the super modifier key working correctly with pynput. There is
a github issue associated with that too:

https://github.com/moses-palmer/pynput/issues/59