Disable and enable are now supported for the python mock.
I added the ttrack integration for both modes too. Hatel is enabled by pressing
all black keys and disabled by pressing 5 white keys. This is because I can
control the keys that can be used in hatel, but I cannot control which keys are
hit in the wild. All black keys is easy to remember and somewhat easy to hear,
so you can know if you are going into hatel or not.